// src/Content/TypeDefinitionFiles/InputLoginContent.d.ts
export type InputLoginContent={
    userName:string,
    profileAppeal:string,
        profileMemo:string,
        name:string,
        address:string,
        telephoneNumber:string,
        mailAddress:string,
        url:string,
        deliverables:string,
        deliverablesTerm:string,
        internship:string,
        selection:string,
        memo:string
}